[Cryptography] rewrite Adler32 by implementing HashAlgorithm
- using the best practise as a foundation
